Removed some white space after merge

+++ efDb/Models/ProductRating.cs
@@ -2,6 +2,6 @@ namespace efDb.Models;
 public class ProductRating{
   public int Id { get; set; }
     public int ProductId { get; set; }
-    public int Rating { get; set; } // Assume rating is between 1 and 5
+    public short Rating { get; set; }
     public Product? Product { get; set; }
 }
